# Faster fill functions in implementations of std.Random