`std.Random.int()` with `std.Random.DefaultPrng` does not generate optimal assembly for all types, at least on x86-64, anything above u56 is not optimized into simple calls to `Xoshiro256.next()`, skipping the serialization step in `Xoshiro256.fill()` and is instead being filled byte by byte.

Comparison of `std.Random.int(u32)` with `std.Random.int(u64)`: <https://zig.godbolt.org/z/czqxoM64b>.

This `std.Random.int()` behaviour also happens with `Sfc64` and `Isaac64`, but other engines that share the same fill function do the same on `std.Random.bytes()`, they generate a result from the `next` function then uses it to fill the buffer a byte at a time.

Calling `std.Random.bytes()` with different engines: <https://zig.godbolt.org/z/8z9dj4eK4>.

Modifying the fill function from `Xoshiro256` such that it doesn't serialize the `next` function result into byte sizes yields these results from poop when benchmarking `std.Random.int(u64)`:

```
$ zig build-exe stdlib.zig -OReleaseFast -fstrip
$ zig build-exe modified.zig -OReleaseFast -fstrip
$ sudo ~/.local/bin/poop "./stdlib" "./modified"
Benchmark 1 (753 runs): ./stdlib
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          6.54ms ±  203us    6.28ms … 7.96ms         74 (10%)        0%
  peak_rss            131KB ±    0       131KB …  131KB          0 ( 0%)        0%
  cpu_cycles            0   ±    0         0   …    0            0 ( 0%)        0%
  instructions          0   ±    0         0   …    0            0 ( 0%)        0%
  cache_references      0   ±    0         0   …    0            0 ( 0%)        0%
  cache_misses          0   ±    0         0   …    0            0 ( 0%)        0%
  branch_misses         0   ±    0         0   …    0            0 ( 0%)        0%
Benchmark 2 (3817 runs): ./modified
  measurement          mean ± σ            min … max           outliers         delta
  wall_time          1.27ms ± 95.6us    1.09ms … 2.43ms         34 ( 1%)        ⚡- 80.5% ±  0.1%
  peak_rss            290KB ± 87.3KB     131KB …  373KB          0 ( 0%)        💩+121.5% ±  4.8%
  cpu_cycles            0   ±    0         0   …    0            0 ( 0%)          -  nan% ± -nan%
  instructions          0   ±    0         0   …    0            0 ( 0%)          -  nan% ± -nan%
  cache_references      0   ±    0         0   …    0            0 ( 0%)          -  nan% ± -nan%
  cache_misses          0   ±    0         0   …    0            0 ( 0%)          -  nan% ± -nan%
  branch_misses         0   ±    0         0   …    0            0 ( 0%)          -  nan% ± -nan%
```

Files for the benchmark: <https://gist.github.com/PauloCampana/b78ed6cf5c42cc323b4a406b7f3f9c38>

And more optimal assembly after the patch: <https://zig.godbolt.org/z/snWYehrWY>