@marler8997 2024-03-15
Tested and confirmed similar benchmark results on linux/windows.  Computers are weird.